# storage: re-introduce dual buffer for iterator bounds