For the `mvcc-range-tombstones` branch. Undoes a change in #79291 and #79744. Will be squashed into the original commit once merged.

---

**storage: re-introduce dual buffer for iterator bounds**

This is necessary to avoid mutating the bounds in use by
`pebble.Iterator`, which can interfere with seek optimizations.

Release note: None

**storage: simplify `pebbleIterator` options comparisons**

Since we now use dual buffers for bounds encoding, we can fully generate
the new `pebble.IterOptions` during `setOptions()` and compare them
directly with the existing options, which simplifies the code.

Release note: None