@erikgrinaker 2022-04-15
Performance diff when writing to a batch (reusing iterators) is negligible, and within error margins (c.f. benchmarks in #79291):

```
name                                        old time/op    new time/op    delta
MVCCGet_Pebble/versions=1/valueSize=8-24      2.60µs ± 4%    2.61µs ± 1%    ~     (p=0.132 n=19+19)
MVCCGet_Pebble/versions=10/valueSize=8-24     3.82µs ± 1%    3.88µs ± 1%  +1.54%  (p=0.000 n=20+19)
MVCCGet_Pebble/versions=100/valueSize=8-24    9.47µs ± 6%    9.59µs ± 3%    ~     (p=0.123 n=19+19)
```